+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/web.php
@@ -19,6 +19,7 @@ use Symfony\Component\HttpKernel\Controller\ArgumentResolver\RequestAttributeVal
 use Symfony\Component\HttpKernel\Controller\ArgumentResolver\RequestValueResolver;
 use Symfony\Component\HttpKernel\Controller\ArgumentResolver\ServiceValueResolver;
 use Symfony\Component\HttpKernel\Controller\ArgumentResolver\SessionValueResolver;
+use Symfony\Component\HttpKernel\Controller\ArgumentResolver\UidValueResolver;
 use Symfony\Component\HttpKernel\Controller\ArgumentResolver\VariadicValueResolver;
 use Symfony\Component\HttpKernel\Controller\ErrorController;
 use Symfony\Component\HttpKernel\ControllerMetadata\ArgumentMetadataFactory;
@@ -48,6 +49,11 @@ return static function (ContainerConfigurator $container) {
         ->set('argument_resolver.backed_enum_resolver', BackedEnumValueResolver::class)
             ->tag('controller.argument_value_resolver', [
                 'priority' => 105, // prior to the RequestAttributeValueResolver
+            ])
+
+        ->set('argument_resolver.uid', UidValueResolver::class)
+            ->tag('controller.argument_value_resolver', [
+                'priority' => 100, // same priority than RequestAttributeValueResolver, but registered before
             ])
 
         ->set('argument_resolver.request_attribute', RequestAttributeValueResolver::class)