+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * Add `BackedEnumValueResolver` to resolve backed enum cases from request attributes in controller arguments
  * Deprecate StreamedResponseListener, it's not needed anymore
  * Add `Profiler::isEnabled()` so collaborating collector services may elect to omit themselves.
+ * Add the `UidValueResolver` argument value resolver
 
 6.0
 ---